backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
flask-cors
flask-psycopg2
argon2-cffi
Flask-Migrate

